doc: makes the library version number fully parametric
The version number is now sourced by the Makefile from the vl/generic.h library header file. webdoc.py now supports a %env:VARIABLE; directive to insert in the documentation the value of an environment variable. The Makefile uses this mechanism to inject the current library version into the documentation.